media-ingestion-service/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
ImageHash
Pillow
sqlalchemy
